refactoring and fix a non-refresh bug
